loolwsd: Broadcast editlock info with viewinfo message
This tells all connected clients about who currently is in
possession of the editlock.

Note that though the editlock information is now embedded in the
viewinfo message, it is not used to turn the editlock on and off
on the client side (yet). Let the 'editlock' message work as it
used to do.

Change-Id: Iba70fb910bcba2cd8e10037f294dbe245c52d83e