Turns out that setting 'blocking' is the right thing to do.
Unfortunately it turns out that POCO's handling of the non-blocking case is
wrong when EWOULDBLOCK is returned from ::send().  This leads to a situation
when part of the data has been sent, but it is impossible to send only the
rest of it, because of too high-level api for the websockets.

We could essentially re-implement the POCO's websockets and use just the lower
levels of POCO - but then it's a question whether it is actually easier to use
the Linux system calls right away :-)

Change-Id: Ied08f68d5037d0ab9ca8266cd20e1084bcebfd14
(cherry picked from commit de9ec63e4f9c0514070016fcaf8aae3772fb7d42)