loleaflet: Don't prune tiles after every tile load
When the connection is slow and tiles are received after a good
amount of lag in between them, one tile load can prune tiles (of
earlier zoom level) for which we have not yet received the tile
response from the server - so there's nothing we have yet to
overlap these tiles completely, leading to grey-ish patches when zooming
in or out.

Pruning these tiles later ensures that we do not remove any tiles
of earlier zoom levels until all the tiles have been loaded.

Change-Id: Ibb03f0c6bf2031741f0b8a4584c1821539587e88
Reviewed-on: https://gerrit.libreoffice.org/35355
Reviewed-by: pranavk <pranavk@collabora.co.uk>
Tested-by: pranavk <pranavk@collabora.co.uk>