Grey-out ("dim") after 10 minutes if idle even while having focus
(The greying-out without focus happens, as before, already after 30
seconds.)

Factor out the dimming code to a separate function _dim. When we have
focus, check once a minute if ten minutes has passed without user
activity. Keep track of when last user activity happened.

There is a bug left: When greyed-out while having focus, you need to
click on the background layer, not in the layer with the 'Inactive
document ...' text, to wake it up.

Change-Id: I19160a82f5de860fc609dad391b168acfba560ce