Slight re-factoring: introduce getParameterValue() function
I started changing the cases for all the parameters to use similar
code as for the 'hash=' one: Check the parameter name with
startsWith() instead of fragile substring() function call where the
length parameter must match the length of a string literal that the
substring() result is compared to.

But then I got bored and gave up. It would be even better to not have
to tediously check for all defined parameter names, but instead just
parse each parameter using the same code that sets the property whose
name we found. Except that we have special cases where the parameter
name in the protocol doesn't match our coresponding property name, and
those that consist of a list of comma-separated values.

Maybe we indeed should just switch to JSON (or some binary
representation of JSON or somesuch).

Change-Id: Ibcdb26d7228b75f6f23c186f0b17cb46a12f3cec