Prevent accidentally having duplicate loolwsds on same port, new approach
Before setting up the socket where we listen for client requests, try
connecting to the same port. If that succeeds, another loolwsd process
is already running and listening on that port. That is obviously
undesirable.

Yes, there is a race condition if multiple loolwsd processes are
started simultaneously and do this check before any of them have
actually created the socket. Live with it. Multiple loolwsd processes
is a problem that happens accidentally for developers only anyway. In
a production environment systemd takes care of having just one, I
hope.

Thanks to Kendy for the idea.

Change-Id: Ifdde83472f9a56e592ec5dc7649dd7706efc2f7c