Revert "loolwsd: improved MessageQueue"
Now we don't get a situation where there would be a tremendous amount of
invalidates & tile render requests piled in the queue, so we can do it
deterministic again.

The only thing that could potentially pile in the queue are the keypresses
events sent from the clients, but that is a different problem anyway.

This reverts commit c326228774e720011b8845f0ee14b4c0a5f1d4f2.

Change-Id: I98e199eab0187bf5f47ce322ac1b1b2e3b976b85