loleaflet: fix readonly UI for spreadsheets
There is still one known problem - for some reason, when a spreasheet is
opened in readonly mode, the ViewRowColumnHeader request sent to server
has height=0 leading to server sending no rows information back to client.
Though the problem gets resolved when one scrolls the spreadsheet
window, but something to fix.

Change-Id: I2aefafd1a93c1bc5a9ad8581edb83018ec9a3f07