nb: write socket buffer to socket safely
Since a socket client can push data into
the socket in a different thread than the one
polling (indeed that's the only possible scenario),
the write buffer must be protected by a lock.

On the other hand, the read buffer is always
invoked from a single thread, the polling. So
it is perfectly safe without locks.

Change-Id: Id0b6a01f8e96124a299810f0aacab9cecd1ff979