Loleaflet - Calc: horizontal scroll bar not updated on changed doc width
For big spreadsheet (> 100k rows) when the doc width is extended the
horizontal scroll-bar is not updated, causing to not be able toscroll
the document up to the whole width.

This was due to the fact that the old doc height was obtained through
L.DomUtil.getStyle, but the precision of the retrieved value is 6
digits only.

Change-Id: I430b08dd4c864cec09b62c61420414ad306da225
(cherry picked from commit 1e24c77c8212c35e4a5fe8f0242c6a531be94c4f)