wsd: take DocBrokers lock before broadcasting diskfull
alertAllUsersInternal does not have the DocBrokersMutex
when called from forkChildren, which was intentionally
removed a while back since it's not normally necessary.
That is, unless we have diskfull situation and need
to broadcast to clients.

We now take the lock explicitly in alertAllUsersInternal
before enumerating DocBrokers to broadcast the notification
to the user, since there is no use case for calling
it under that lock now.

Change-Id: I4f7cccb1112c17091d1039441af4c84d5d13d059
Reviewed-on: https://gerrit.libreoffice.org/36601
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>
Tested-by: Ashod Nakashian <ashnakash@gmail.com>
(cherry picked from commit 6023fdc3c70a2d2aad2c98c8e739f19d7a630e8b)
Reviewed-on: https://gerrit.libreoffice.org/36611
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Jan Holesovsky <kendy@collabora.com>