loleaflet: Allow users to repair at all times
Since during an edit conflict the undo/redo
buttons are dissabled (for the user who can't
undo/redo anymore) disabling the repair button
robs them of the ability to repair the document
to an earlier state (i.e. before the conflict).

This enables the repair button as soon as either
the undo or redo is enabled, whereupon it remains
enabled for the duration of the session.

Change-Id: Ibaf257155f721e2cef596e965d8812d381ab3a5b
Reviewed-on: https://gerrit.libreoffice.org/38942
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>
Tested-by: Ashod Nakashian <ashnakash@gmail.com>
(cherry picked from commit 323e2789b9e82e0d600d39d0697d40f6014f54d7)
Reviewed-on: https://gerrit.libreoffice.org/38943
Reviewed-by: pranavk <pranavk@collabora.co.uk>
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Jan Holesovsky <kendy@collabora.com>