lookitconfig.xcu: Set a work path which is valid in chroot
Since the variables are set outside of chroot, workdir directory is
wrongly picked up as /home/username/Documents or something else. But it
is not guarranteed that this directory would exist in the chroot. Any
attempt to access such an invalid directory, eg. during filepicker
button in hyperlink dialog, freezes the dialog.

Change-Id: I83bfc5787a4f3bd66d8abd30b6204cb1e74e5287