Add handling of the SetCellColor proof-of-concept postMessage event
When we receive such a message, we forward it to the corresponding
Python script (from core's scripting/examples/python/SetCellColor.py).
We also save the event's source so that we can send the return value
of the script back to its window.

[Later addition, when pushing upstream: This commit will be reverted
at least partially by later commits. I didn't bother with squashing
them all into one. So sue me. Also, it might be of interest to see in
the commit log how my ideas for how to implement this stuff evolved.]

Change-Id: I1f54a90000937d6c9b91fb15cdfde329223cf8db