wsd: refactor diskpace and session limit checks
Checks are now done after adding new sessions
to DocBroker rather than before, so the current
session being added doesn't need special handling.

The checks for diskspace is separated from that
of the number of sessions as they are unrelated.

Also, no reason to do the checks for convert-to
requests, since these don't have interactive
clients, rather the connections are closed after
the conversion.

Change-Id: Idc50cd38263e6779acdeed72d5eb876a3228c96e
Reviewed-on: https://gerrit.libreoffice.org/52418
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Michael Meeks <michael.meeks@collabora.com>