loleaflet: Keep the hidden input width 1x1 to keep safari happy
It's not possible otherwise to focus the <input> html element in Safari.
Unfortunately, making the input field 1x1 has the disadvantage that on
IE 11, a very small blinking artifact appears on bigger font size. There
doesn't seem to be any other solution to hide the cursor on IE without
messing up with IME candidate window position.

Without this patch, typing doesn't work on Safari at all which is much
much bigger of an issue than some artifact appearing on IE.

Change-Id: Ib618a96ba2ec3aa621ed448ffbf410113d8975b8
Reviewed-on: https://gerrit.libreoffice.org/51579
Reviewed-by: Aron Budea <aron.budea@collabora.com>
Tested-by: Aron Budea <aron.budea@collabora.com>