Revert all related loleaflet IME changes
... that can potentially affect the html/css layout. We have a problem
in the front-end where map gets incorrectly positioned due to the moving
candidate window. With this revert, the candidate window is thrown back
at where it was (to the top-left) and the problem we are facing (as
mentioned above) dissappears as well. Let's fix that problem first
before re-revert it.

Change-Id: I1b629de51fb5f82dceb007105eb4c3ee1d411d22
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Tested-by: Andras Timar <andras.timar@collabora.com>