wsd: don't log binary data from sockets as text
Instead of dumping arbitrary binary data from sockets
we now dump the first 96 bytes as hex.

The dumpHex utility function has been capitalized for
this purpose, with some improvements.

It is now possible to cap the number of bytes dumped
as hex.
An extra line of hex dump was outputed, now omitted.
When the data ends, we don't output hex values and
we also don't output '.' either, instead we output
blank.

Change-Id: I0d3ac22f27cd8a7f6b5951653f061f52b8333dc5