Use a bigger timeout for waiting tileprocessed message
The main purpose of this time-out is to avoid waiting forever for lost
tile messages, but since it rare to loose them we can use bigger value,
so we can avoid to send new tiles to a slow network.

Note that the used time stamp does not mean the time when the tile
actually send to the client, but the time when it gets to the sender
queue.

Change-Id: I230d85c38b3a5dafd195851d0cf4caac23149e3e
(cherry picked from commit b623aca57ea47c785ab70ff6ccec58acbff63d75)
Reviewed-on: https://gerrit.libreoffice.org/60879
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Jan Holesovsky <kendy@collabora.com>
(cherry picked from commit 7253dd1499db54550a4e4e887820027c9c8a4c31)