Avoid rendering / sending the same tile twice
When we're doing the prerendering we also need to register
a tileBeingRendered object, so we know that the given tile
will arrive soon. In earlier version of the code, rendering
and sending of the tile was not separated in time, but now it is,
so we need create a tileBeingRendered object even if we don't
subscribe the client to it yet.

(cherry picked from commit 5260eae05f89056d4ffc1bdfad1927c3c5f9706d)

Change-Id: I1374c22e5ffebe1d6fcc6e37261ddcedeb9066ec
Reviewed-on: https://gerrit.libreoffice.org/61124
Reviewed-by: Jan Holesovsky <kendy@collabora.com>
Tested-by: Jan Holesovsky <kendy@collabora.com>
(cherry picked from commit c2d800bd345a14c006697e4fd8a75fcadf36557c)