android chrome: Stop composing when we get an actual key event.
So that it is possible to press Enter while still in composition.
It is necessary to stop the composing itself, because otherwise the word
gets duplicated on the new line as soon as the user presses a key.

Change-Id: I78951a423715e71533f1a73d5bbe3b0f0f05e6cd
Reviewed-on: https://gerrit.libreoffice.org/61334
Reviewed-by: Aron Budea <aron.budea@collabora.com>
Tested-by: Aron Budea <aron.budea@collabora.com>
(cherry picked from commit f42e0bb4d1c99c1d229a638e59061ef7985d01e8)